Group context and hook creation so that it is exportable.
